pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "x-collector"
version = "0.1.0"
description = "X/Twitter data collection tool for OpenClaw"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "CoinSummer Research" }
]
keywords = ["x", "twitter", "collector", "openclaw", "tweets"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.0",
]

[project.scripts]
x-collector = "x_collector.cli:main"

[project.urls]
Homepage = "https://github.com/coinsummer/x-collector"
Documentation = "https://github.com/coinsummer/x-collector#readme"

[tool.hatch.build.targets.wheel]
packages = ["x_collector"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
